Hide activateNatReprCoercionBackdoor from external use
